+++ src/services/add-query-to-queue.ts
@@ -6,9 +6,9 @@ import {TYPES} from '../types.js';
 import GetSongs from '../services/get-songs.js';
 import {SongMetadata, STATUS} from './player.js';
 import PlayerManager from '../managers/player.js';
-import {prisma} from '../utils/db.js';
 import {buildPlayingMessageEmbed} from '../utils/build-embed.js';
 import {getMemberVoiceChannel, getMostPopularVoiceChannel} from '../utils/channels.js';
+import {getGuildSettings} from '../utils/get-guild-settings';
 
 @injectable()
 export default class AddQueryToQueue {
@@ -34,11 +34,7 @@ export default class AddQueryToQueue {
 
     const [targetVoiceChannel] = getMemberVoiceChannel(interaction.member as GuildMember) ?? getMostPopularVoiceChannel(interaction.guild!);
 
-    const settings = await prisma.setting.findUnique({where: {guildId}});
-
-    if (!settings) {
-      throw new Error('Could not find settings for guild');
-    }
+    const settings = await getGuildSettings(guildId);
 
     const {playlistLimit} = settings;