+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- Create the guild settings when not found instead of returning an error
 - Add temporary workaround to avoid VoiceConnection being stuck in signalling state
 
 ## [2.2.0] - 2023-02-26